perl: Improve and update module RPDEPENDS
The perl module RDEPENDS needed refreshing so I updated it with the script. I
also found a ton of issues with missing test2 module dependencies so I've tweaked
the generator script to improve those. For some reason they were previously excluded
but I can't see anything wrong with the generated dependencies so I've added them. A
number of manual dependencies were still needed in addition.

I did notice an issue with trailing ")" characters so the sed is updated to handle
that (it did already handle "(").

This fixes issues seen in the dependecies for libxml-parser-perl's ptest package.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>